refactor: extract ReviewSummaryIssueGroup.swift from ReviewWorkspaceTypes.swift
Splits the `ReviewSummaryIssueGroup` struct (category grouping for
the review summary tab) out of ReviewWorkspaceTypes.swift so that file
stays focused on tab + timeline entry types.

Byte-identical move of lines 57-73. Both files import Foundation.

ReviewWorkspaceTypes.swift: 73 → 55 lines. Tests: 667.
