+++ Sources/BugNarrator/Utilities/ReviewWorkspaceTypes.swift
@@ -53,21 +53,3 @@ enum ReviewWorkspaceTimelineEntryKind: Equatable {
         }
     }
 }
-
-struct ReviewSummaryIssueGroup: Equatable {
-    let category: ExtractedIssueCategory
-    let issues: [ExtractedIssue]
-
-    var title: String {
-        switch category {
-        case .bug:
-            return "Bugs"
-        case .uxIssue:
-            return "UX Issues"
-        case .enhancement:
-            return "Enhancements"
-        case .followUp:
-            return "Questions / Follow-ups"
-        }
-    }
-}
